Reduce the response that gameEngine.handleInput returns.
The caller has access to the gameState object, so the engine should leave it up to the caller to return whatever it needs to from gamestate.

However, we want to guarantee that the events are flushed, so the gamestate does that and returns the events generated during the command execution.